+++ lib/src/cmd/onvif_debug_command.dart
@@ -1,9 +1,14 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:archive/archive_io.dart';
 import 'package:args/command_runner.dart';
 import 'package:cli_spin/cli_spin.dart';
 import 'package:easy_onvif/onvif.dart';
+import 'package:easy_onvif/shared.dart';
+import 'package:easy_onvif/src/model/device_management/user.dart';
+import 'package:easy_onvif/src/model/media1/stream_setup.dart';
+import 'package:easy_onvif/src/model/media1/transport.dart';
 import 'package:easy_onvif/util.dart';
 import 'package:loggy/loggy.dart';
 import 'package:path/path.dart' as p;
@@ -64,6 +69,8 @@ class OnvifDebugCommand extends Command with UiLoggy {
       spinner: CliSpinners.line,
     ).start(); // Chai
 
+    final random = Random().nextInt(900000) + 100000;
+
     Timer(Duration(milliseconds: 1000), () {
       // Change spinner color
       spinner.color = CliSpinnerColor.yellow;
@@ -72,7 +79,15 @@ class OnvifDebugCommand extends Command with UiLoggy {
       spinner.text = 'Still loading, please wait...';
     });
 
+    //
     // get device management information
+    //
+
+    try {
+      await onvif.deviceManagement.getCapabilities();
+    } catch (e) {
+      loggy.error('getCapabilities');
+    }
 
     try {
       await onvif.deviceManagement.getDeviceInformation();
@@ -123,6 +138,18 @@ class OnvifDebugCommand extends Command with UiLoggy {
     }
 
     try {
+      await onvif.deviceManagement.getSystemDateAndTime();
+    } catch (e) {
+      loggy.error('getSystemDateAndTime');
+    }
+
+    try {
+      await onvif.deviceManagement.getSystemLog('System');
+    } catch (e) {
+      loggy.error('getSystemDateAndTime');
+    }
+
+    try {
       await onvif.deviceManagement.getSystemUris();
     } catch (e) {
       loggy.error('getSystemUris');
@@ -134,7 +161,26 @@ class OnvifDebugCommand extends Command with UiLoggy {
       loggy.error('getUsers');
     }
 
+    try {
+      await onvif.deviceManagement.createUsers([
+        User(
+            username: 'deleteMe_$random',
+            password: 'deleteMe',
+            userLevel: UserLevel.user)
+      ]);
+    } catch (e) {
+      loggy.error('getUsers');
+    }
+
+    try {
+      await onvif.deviceManagement.deleteUsers(['deleteMe_$random']);
+    } catch (e) {
+      loggy.error('getUsers');
+    }
+
+    //
     // get media1 information
+    //
 
     try {
       await onvif.media.media1.getAudioSources();
@@ -142,20 +188,44 @@ class OnvifDebugCommand extends Command with UiLoggy {
       loggy.error('getAudioSources');
     }
 
-    // await onvif.media.media1.getMetadataConfiguration(configurationToken)
-
     try {
-      await onvif.media.media1.getMetadataConfigurations();
+      var metadataConfigurations =
+          await onvif.media.media1.getMetadataConfigurations();
+
+      await onvif.media.media1
+          .getMetadataConfiguration(metadataConfigurations.first.token);
     } catch (e) {
       loggy.error('getMetadataConfigurations');
+      loggy.error('getMetadataConfiguration');
     }
 
-    // await onvif.media.media1.getProfile(profileToken);
-
     try {
-      await onvif.media.media1.getProfiles();
+      var profiles = await onvif.media.media1.getProfiles();
+
+      await onvif.media.media1.getProfile(profiles.first.token);
+
+      await onvif.media.media1.getSnapshotUri(profiles.first.token);
+
+      await onvif.media.media1.getStreamUri(
+        profiles.first.token,
+        streamSetup: StreamSetup(
+          stream: 'RTP-Unicast',
+          transport: Transport(
+            protocol: 'RTSP',
+          ),
+        ),
+      );
+
+      await onvif.media.media1.startMulticastStreaming(profiles.first.token);
+
+      await onvif.media.media1.stopMulticastStreaming(profiles.first.token);
     } catch (e) {
       loggy.error('getProfiles');
+      loggy.error('getProfile');
+      loggy.error('getSnapshotUri');
+      loggy.error('getStreamUri');
+      loggy.error('startMulticastStreaming');
+      loggy.error('stopMulticastStreaming');
     }
 
     try {
@@ -164,18 +234,15 @@ class OnvifDebugCommand extends Command with UiLoggy {
       loggy.error('getServiceCapabilities');
     }
 
-    // await onvif.media.media1.getSnapshotUri(profileToken);
-
-    // await onvif.media.media1
-    //     .getStreamUri(profileToken, streamSetup: streamSetup);
-
     try {
       await onvif.media.media1.getVideoSources();
     } catch (e) {
       loggy.error('getVideoSources');
     }
 
+    //
     // get media2 information
+    //
 
     try {
       await onvif.media.media2.getMetadataConfigurationOptions();
@@ -190,9 +257,43 @@ class OnvifDebugCommand extends Command with UiLoggy {
     }
 
     try {
-      await onvif.media.media2.getProfiles();
+      var profiles = await onvif.media.media2.getProfiles();
+
+      var videoEncoderConfiguration =
+          profiles.first.configurations?.videoEncoderConfiguration;
+
+      var videoSourceConfiguration =
+          profiles.first.configurations?.videoSourceConfiguration;
+
+      await onvif.media.media2.getSnapshotUri(profiles.first.token);
+
+      await onvif.media.media2.getStreamUri(
+        profiles.first.token,
+        protocol: 'RTSP',
+        streamType: 'RTP-Unicast',
+      );
+
+      await onvif.media.media2.startMulticastStreaming(profiles.first.token);
+
+      await onvif.media.media2.stopMulticastStreaming(profiles.first.token);
+
+      if (videoEncoderConfiguration != null) {
+        await onvif.media.media2
+            .getVideoEncoderInstances(videoEncoderConfiguration.token);
+      }
+
+      if (videoSourceConfiguration != null) {
+        await onvif.media.media2.getVideoSourceConfigurationOptions(
+            configurationToken: videoSourceConfiguration.token);
+      }
     } catch (e) {
       loggy.error('getProfiles');
+      loggy.error('getSnapshotUri');
+      loggy.error('getStreamUri');
+      loggy.error('startMulticastStreaming');
+      loggy.error('stopMulticastStreaming');
+      loggy.error('getVideoEncoderInstances');
+      loggy.error('getVideoSourceConfigurationOptions');
     }
 
     try {
@@ -201,7 +302,53 @@ class OnvifDebugCommand extends Command with UiLoggy {
       loggy.error('getServiceCapabilities');
     }
 
+    try {
+      await onvif.media.media2.getVideoEncoderConfigurations();
+    } catch (e) {
+      loggy.error('getVideoEncoderConfigurations');
+    }
+
+    try {
+      await onvif.media.media2.getVideoSourceConfigurationOptions();
+    } catch (e) {
+      loggy.error('getServiceCapabilities');
+    }
+
+    //
     // get ptz information
+    //
+
+    try {
+      var profiles = await onvif.media.media1.getProfiles();
+
+      var profileToken = profiles.first.token;
+
+      await onvif.ptz.absoluteMove(
+        profileToken,
+        position: PtzVector(
+          panTilt: Vector2D(x: 0, y: 0),
+          zoom: Vector1D(x: 0),
+        ),
+      );
+
+      await onvif.ptz.continuousMove(
+        profileToken,
+        velocity: PtzSpeed(
+          panTilt: Vector2D(x: 0.1, y: 0.1),
+          zoom: Vector1D(x: 0),
+        ),
+      );
+
+      var configurations =
+          await onvif.ptz.getCompatibleConfigurations(profileToken);
+
+      await onvif.ptz.getConfigurationOptions(configurations.first.token);
+    } catch (e) {
+      loggy.error('absoluteMove');
+      loggy.error('continuousMove');
+      loggy.error('getCompatibleConfigurations');
+      loggy.error('getConfigurationOptions');
+    }
 
     try {
       await onvif.ptz.getServiceCapabilities();
@@ -215,9 +362,9 @@ class OnvifDebugCommand extends Command with UiLoggy {
       loggy.error('getConfigurations');
     }
 
-    // await onvif.ptz.getCompatibleConfigurations(profileToken);
-
-    // await onvif.ptz.getConfigurationOptions(configurationToken);
+    //
+    // done with checks
+    //
 
     final archive = Archive();