+++ lib/src/model/media1/video_source.g.dart
@@ -11,7 +11,7 @@ VideoSource _$VideoSourceFromJson(Map<String, dynamic> json) => VideoSource(
       frameRate:
           OnvifUtil.mappedToString(json['Framerate'] as Map<String, dynamic>),
       resolution:
-          Resolution.fromJson(json['Resolution'] as Map<String, dynamic>),
+          VideoResolution.fromJson(json['Resolution'] as Map<String, dynamic>),
       imaging: json['Imaging'] == null
           ? null
           : Imaging.fromJson(json['Imaging'] as Map<String, dynamic>),